requirements.txt
spacy>=3.5
watchdog>=4.0
pdfplumber>=0.11
python-docx>=1.0
